move mifare stuff to separate folder
