+++ client/cmdhfmf.h
@@ -11,8 +11,9 @@
 #ifndef CMDHFMF_H__
 #define CMDHFMF_H__
 
-#include "mifaredefault.h"
+#include "mifare/mifaredefault.h"
 
+extern int CmdHFMF(const char *Cmd);
 extern int CmdHFMF(const char *Cmd);
 
 extern int CmdHF14AMfDbg(const char* cmd);
