+++ client/cmdhfmfp.h
@@ -10,7 +10,7 @@
 #ifndef CMDHFMFP_H__
 #define CMDHFMFP_H__
 
-#include "mifaredefault.h"
+#include "mifare/mifaredefault.h"
 
 extern int CmdHFMFP(const char *Cmd);
